+++ R/00_utils.R
@@ -121,6 +121,3 @@ with_openxml_unquote <- function(x) {
 with_html_quotes <- function(x) {
   paste("```{=html}", x, "```\n\n", sep = "\n")
 }
-with_html_unquotes <- function(...) {
-  paste("\n\n```\n", ..., "\n```{=html}\n\n", sep = "\n")
-}