+++ NEWS.md
@@ -4,10 +4,28 @@
 
 - `headers_flextable_at_bkm()` and `footers_flextable_at_bkm()` are defunct.
 - `flextable_to_rmd()` is now using `knit_child()` for safer usage from `for`
-loops or `if` statements.
-- Add explanation about caption limitations in the manual of functions 
-`save_as_image()` and `ph_with.flextable()`.
-- Deprecate `as_raster()` since `gen_grob()` is easier to use and render nicer.
+  loops or `if` statements.
+- Add explanation about caption limitations in the manual of functions
+  `save_as_image()` and `ph_with.flextable()`.
+- Deprecate `as_raster()` since `gen_grob()` is easier to use and render
+  nicer.
+- BREAKING CHANGE: in `align()`, the default argument value for `align` is now
+  `"left"`, rather than `c("left", "center", "right", "justify")`. This
+  returns the default value to how it was in older versions of {flextable}.
+    - in `align()`, use of the old default `align` argument could cause an
+      error if the number of columns being adjusted was not a multiple of 4.
+    - The documentation specified that `align` had to be a single value, when
+      it could actually accept multiple values. This is why a default value of
+      `c("left", "center", "right", "justify")`, was problematic. This
+      documentation has now been updated and new examples included in the
+      documentation.
+    - The default `align` argument will now apply left alignment to all
+      columns in the body.
+    - If the user specifies an alignment that is invalid, a error will be
+      displayed.
+    - The `path` argument now has a signature of `part = c("body", "header",
+      "footer", "all")`, but because only a single value can be selected, it
+      will pick `"body"` by default, as before.
 
 ## Issues